modulize components
